requirements-dev.txt
pylint==3.0.3
coverage==7.4.1
google-api-python-client==2.116.0
google-auth==2.27.0
pandas==2.2.0
numpy==1.26.3